marketing_analytics_backend/requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
apscheduler==3.10.4
pyjwt==2.8.0
pandas==2.1.1
gunicorn==21.2.0
pytest==7.4.3
supabase==2.0.3
